+++ x/wrkchain/genesis.go
@@ -2,7 +2,9 @@ package wrkchain
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/spf13/viper"
 	abci "github.com/tendermint/tendermint/abci/types"
+	undtypes "github.com/unification-com/mainchain/types"
 )
 
 func InitGenesis(ctx sdk.Context, keeper Keeper, data GenesisState) []abci.ValidatorUpdate {
@@ -23,17 +25,23 @@ func InitGenesis(ctx sdk.Context, keeper Keeper, data GenesisState) []abci.Valid
 		for _, block := range record.WrkChainBlocks {
 			blk := WrkChainBlock{
 				WrkChainID: wrkChain.WrkChainID,
-				Height: block.Height,
-				BlockHash: block.BlockHash,
+				Height:     block.Height,
+				BlockHash:  block.BlockHash,
 				ParentHash: block.ParentHash,
-				Hash1: block.Hash1,
-				Hash2: block.Hash2,
-				Hash3: block.Hash3,
+				Hash1:      block.Hash1,
+				Hash2:      block.Hash2,
+				Hash3:      block.Hash3,
 				SubmitTime: block.SubmitTime,
-				Owner: wrkChain.Owner,
+				Owner:      wrkChain.Owner,
 			}
 			//logger.Info("Registering Block for WRKChain", "wc_id", wrkChain.WrkChainID, "h", block.Height)
 			err = keeper.SetWrkChainBlock(ctx, blk)
+			if err != nil {
+				panic(err)
+			}
+
+			// also update NumBlocks!
+			err = keeper.SetNumBlocks(ctx, wrkChain.WrkChainID)
 			if err != nil {
 				panic(err)
 			}
@@ -48,6 +56,7 @@ func ExportGenesis(ctx sdk.Context, k Keeper) GenesisState {
 	initialWrkChainID, _ := k.GetHighestWrkChainID(ctx)
 
 	wrkChains := k.GetAllWrkChains(ctx)
+	exportWrkChainDataIds := viper.GetIntSlice(undtypes.FlagExportIncludeWrkchainData)
 
 	if len(wrkChains) == 0 {
 		return GenesisState{
@@ -58,10 +67,25 @@ func ExportGenesis(ctx sdk.Context, k Keeper) GenesisState {
 	}
 
 	for _, wc := range wrkChains {
-		wrkchainId := wc.WrkChainID
-		blockHashList := k.GetAllWrkChainBlockHashesForGenesisExport(ctx, wrkchainId)
+		exportData := false
+		for _, expWrkChainId := range exportWrkChainDataIds {
+			if uint64(expWrkChainId) == wc.WrkChainID {
+				exportData = true
+			}
+		}
 
-		records = append(records, WrkChainExport{WrkChain: wc, WrkChainBlocks: blockHashList})
+		if exportData {
+			wrkchainId := wc.WrkChainID
+			blockHashList := k.GetAllWrkChainBlockHashesForGenesisExport(ctx, wrkchainId)
+			if blockHashList == nil {
+				blockHashList = WrkChainBlocksGenesisExport{}
+			}
+			records = append(records, WrkChainExport{WrkChain: wc, WrkChainBlocks: blockHashList})
+		} else {
+			wc.LastBlock = 0
+			wc.NumberBlocks = 0
+			records = append(records, WrkChainExport{WrkChain: wc, WrkChainBlocks: WrkChainBlocksGenesisExport{}})
+		}
 	}
 	return GenesisState{
 		Params:             params,